Support role-based profiles
AWS credential profiles can be role-based, meaning they use the Security
Token Service (STS) to assume a role, which in turn returns a set of
short-lived temporary credentials with that role's permissions.

This change introduces the `getCredentialsProvider()` method.  It uses
`ProfilesConfigFile` from AWS to examine the profile.  If it's a
role-based profile, it returns an
`STSAssumeRoleSessionCredentialsProvider`; otherwise it uses the same
`ProfileCredentialsProvider` used previously.

For non-role-based profiles, it should continue to work as it did
previously.  For role-based profiles, it now correctly assumes the role
from the profile and communicates with AWS services.